Pulling RawKeepInterner out of KeepInterner to break circular dependency, simplify twitter imports
